Moved all files into `src`
