+++ example/lib/detailscreen.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:showcaseview/showcase_widget.dart';
 import 'package:showcaseview/showcaseview.dart';
 
 class Detail extends StatefulWidget {
